Changed release level narrative for better conformance with ASF policy.
Added default release level so developers don't have to be
concerned with release policy during configuration.  This is
implemented as a new ./RELEASE_LEVEL file.

Added platform-specific JDK header file to '-I' include path.
This is queried from the user from a list of valid subdirectories
to JAVA_HOME/include


git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/trunk@327202 13f79535-47bb-0310-9956-ffa450edef68