Changed FILL_INFO_DUP[012] defines to new naming method FILL_INFO_MEM_ALIGN[012] to be consistent with the code changes.
Moved PTR_DST_XXX_AI() and PTR_XXX_AI() macros out
to 'attribute.h' for sharing with other areas of the
code.

Renamed some 'src' and 'dst' pointers to 'pev' and 'pevpma'
to be more descriptive as a pointer to element value
and pointer to element value pair, memory aligned,
respectively.

Renamed some 'src' and 'dst' anpointers to 'pn' and 'pnma'
to be more descriptive as a pointer to annotation and
pointer to annotation, memory aligned, respectively.

Renamed some 'src' and 'dst' pointers as 'pai' and 'paima'
or 'ppaima' to be more descriptive as a pointer to
attribute_info and [pointer to] pointer to attribute_info,
memory aligned, respectively.


git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/trunk@418266 13f79535-47bb-0310-9956-ffa450edef68