Move exception handling to beginning of methods.
Make sure all exceptions are using Integer.valueOf(...) for int arguments.

In case where the exception moves lead to restructuring of if else
statements, I also re-ordered some of the statements to make the more
complicated conditions/cases come later.  This should actually
marginally help optimize for the (more typical?) cases of adding items
to the beginning and/or end of the list.


git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/java/trunk@982250 13f79535-47bb-0310-9956-ffa450edef68